Moved addtl utils, made install reuse async cleanup, reverted CHANGELOG
