+++ lib/app-dependencies.js
@@ -2,7 +2,7 @@ const latestVersion = require('latest-version');
 const getPackageJson = require('package-json');
 const logger = require('@blackbaud/skyux-logger');
 
-const sortUtils = require('../lib/sort-utils');
+const sortUtils = require('../lib/utils/sort-utils');
 
 /**
  * Alphabetizes dependencies, similar to the behavior of `npm install <package> --save`.
