+++ lib/install.js
@@ -1,6 +1,6 @@
-const fs = require('fs-extra');
 const logger = require('@blackbaud/skyux-logger');
 const npmInstall = require('./utils/npm-install');
+const cleanup = require('../lib/cleanup');
 
 const settings = {};
 
@@ -8,30 +8,13 @@ if (logger.logLevel === 'verbose') {
   settings.stdio = 'inherit';
 }
 
-function remove(target) {
-  const message = logger.promise(`Remove ${target}.`);
-
-  return fs.remove(target)
-    .then(() => message.succeed())
-    .catch(err => {
-      message.fail();
-      logger.error(err);
-    });
-}
-
-function removeNodeModules() {
-  return remove('node_modules');
-}
-
-function removePackageLock() {
-  return remove('package-lock.json');
-}
-
-function install() {
-  return removeNodeModules()
-    .then(() => removePackageLock())
-    .then(() => npmInstall(settings))
-    .catch(logger.error);
+async function install() {
+  try {
+    await cleanup.deleteDependencies();
+    await npmInstall(settings);
+  } catch(e) {
+    logger.error(e);
+  }
 }
 
 module.exports = install;
