+++ lib/upgrade.js
@@ -1,7 +1,7 @@
 const logger = require('@blackbaud/skyux-logger');
 
 const appDependencies = require('../lib/app-dependencies');
-const jsonUtils = require('../lib/json-utils');
+const jsonUtils = require('../lib/utils/json-utils');
 const cleanup = require('../lib/cleanup');
 
 function isReferenced(packageName, packageJson) {
